requirements.txt
pytorch==1.12.1
tensorflow==2.10.0
scikit-learn==1.1.2
pytorch-lightning==1.7.7
transformers==4.23.1
emoji==2.1.0
wordsegment==1.3.1
sacrebleu
sentencepiece
seaborn
matplotlib